requirements.txt
APScheduler==3.6.3
asn1crypto==0.24.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.5
chardet==3.0.4
charset-normalizer==2.0.7
colorzero==1.1
cryptography==3.4.7
entrypoints==0.3
gpiozero==1.5.1
idna==3.3
keyring==17.1.1
keyrings.alt==3.1.1
olefile==0.46
pexpect==4.6.0
Pillow==5.4.1
pycparser==2.20
pycrypto==2.6.1
pyftpdlib==1.5.6
PyGObject==3.30.4
python-apt==1.8.4.3
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
reportlab==3.5.13
requests==2.26.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.16.0
spidev==3.4
ssh-import-id==5.7
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
